+++ fwd.h
@@ -240,7 +240,7 @@ extern DLI int                  ici_re_nbra;
 extern DLI volatile int         ici_aborted;            /* See exec.c */
 extern DLI int                  ici_dont_record_line_nums;      /* See lex.c */
 extern DLI char                 *ici_buf;                       /* See buf.h */
-extern DLI int                  ici_bufz;                       /* See buf.h */
+extern DLI size_t               ici_bufz;                       /* See buf.h */
 extern DLI ici_ftype_t          ici_stdio_ftype;
 extern DLI ici_ftype_t          ici_popen_ftype;
 extern DLI ici_mark_t           ici_o_mark;
